+++ src/main/java/com/todo/service/UserService.java
@@ -3,7 +3,7 @@ package com.todo.service;
 import com.todo.dto.UserRequestDTO;
 import com.todo.dto.UserResponseDTO;
 import com.todo.entity.User;
-import jakarta.persistence.EntityNotFoundException;
+import com.todo.exception.UserNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,68 +11,51 @@ import com.todo.repository.UserRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 
-
 @Service
 @Transactional
 public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    //private int id = 0;
-
     public UserResponseDTO createUser(UserRequestDTO userRequestDTO) {
         User user = new User();
-
-        //user.setId(++id);
         user.setName(userRequestDTO.getName());
         user.setPassword(userRequestDTO.getPassword());
         user.setEmail(userRequestDTO.getEmail());
 
-        User save = userRepository.save(user);
+        User savedUser = userRepository.save(user);
 
         UserResponseDTO userResponseDTO = new UserResponseDTO();
-
-        userResponseDTO.setId(save.getId());//kullanıcı id sini otomatik alır
-        userResponseDTO.setName(save.getName());
-        userResponseDTO.setEmail(save.getEmail());
+        userResponseDTO.setId(savedUser.getId());
+        userResponseDTO.setName(savedUser.getName());
+        userResponseDTO.setEmail(savedUser.getEmail());
 
         return userResponseDTO;
     }
 
     public List<UserResponseDTO> findAll() {
-        return userRepository.findAll().stream().map(user -> { //tüm kullanıcıları veritabanından alır
+        return userRepository.findAll().stream().map(user -> {
             UserResponseDTO userResponseDTO = new UserResponseDTO();
             userResponseDTO.setName(user.getName());
             userResponseDTO.setEmail(user.getEmail());
             userResponseDTO.setId(user.getId());
             return userResponseDTO;
-        }).collect(Collectors.toList());  //streami listeye döndürür.
+        }).collect(Collectors.toList());
     }
 
     public UserResponseDTO findById(Long id) {
-        User user = userRepository.findById(id).get();
+        User user = userRepository.findById(id).orElseThrow(() ->new UserNotFoundException("User with id " + id + " not found"));
 
         UserResponseDTO userResponseDTO = new UserResponseDTO();
         userResponseDTO.setName(user.getName());
         userResponseDTO.setEmail(user.getEmail());
         userResponseDTO.setId(user.getId());
 
-//        userResponseDTO.setTasks(user.getTasks().stream().map(task -> {
-//            TaskResponseDTO taskResponseDTO = new TaskResponseDTO();
-//            taskResponseDTO.setId(task.getId());
-//            taskResponseDTO.setTitle(task.getTitle());
-//            taskResponseDTO.setDescription(task.getDescription());
-//            taskResponseDTO.setStatus(task.getStatus());
-//            taskResponseDTO.setPriority(task.getPriority());
-//            return taskResponseDTO;
-//
-//        }).collect(Collectors.toList()));
-
         return userResponseDTO;
     }
 
     public UserResponseDTO updateUser(Long id, UserRequestDTO userRequestDTO) {
-        User user = userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
+        User user = userRepository.findById(id).orElseThrow(() -> new UserNotFoundException("User with id"+ id +" not found"));
         user.setName(userRequestDTO.getName());
         user.setEmail(userRequestDTO.getEmail());
         user.setPassword(userRequestDTO.getPassword());
@@ -83,14 +66,16 @@ public class UserService {
         userResponseDTO.setId(user.getId());
         userResponseDTO.setName(user.getName());
         userResponseDTO.setEmail(user.getEmail());
-        //user.setPassword(userRequestDTO.getPassword());
 
         return userResponseDTO;
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User with id " + id + " not found");
+        }
         userRepository.deleteById(id);
-        // userResponseDTO.setMesaj("başarılı silme işlemi");
+        System.out.println("Deleted user with id " + id);
     }
 }
